Fix a common thunk case (splitting it from redux-json-api thunk)
