+++ src/common/store/tabs/tab/api/action-aliases.ts
@@ -5,12 +5,26 @@ import {
   deleteResource as originalDeleteResource,
   requireResource as originalRequireResource,
 } from 'redux-json-api';
-import { aliasActionToTabMarkedThunk } from 'common/store/action-utils';
+import { markInThunkActionWithTabId, retrieveAnyActionTab } from '../../action-tab';
+
+// A redux-json-api version of common/action-utils converter
+// Amend getState to return current tab state (since redux-json-api assumes its state is preserved in the root)
+export function reduxJsonApiAliasActionToTabMarkedThunk(originalThunk) {
+  return (aliasAction) => {
+    return (dispatch, getState) => {
+      const tabId = retrieveAnyActionTab(aliasAction);
+      // attach the tabId to all actions dispatched within the thunk
+      const newDispatch = action => dispatch(markInThunkActionWithTabId(action, tabId));
+      const newGetState = () => getState().tabs[tabId];
+      return originalThunk(...aliasAction.args)(newDispatch, newGetState);
+    };
+  };
+}
 
 export const reduxJsonApiAliases = {
-  createResource: aliasActionToTabMarkedThunk(originalcreateResource),
-  readEndpoint: aliasActionToTabMarkedThunk(originalReadEndpoint),
-  updateResource: aliasActionToTabMarkedThunk(originalUpdateResource),
-  deleteResource: aliasActionToTabMarkedThunk(originalDeleteResource),
-  requireResource: aliasActionToTabMarkedThunk(originalRequireResource),
+  createResource: reduxJsonApiAliasActionToTabMarkedThunk(originalcreateResource),
+  readEndpoint: reduxJsonApiAliasActionToTabMarkedThunk(originalReadEndpoint),
+  updateResource: reduxJsonApiAliasActionToTabMarkedThunk(originalUpdateResource),
+  deleteResource: reduxJsonApiAliasActionToTabMarkedThunk(originalDeleteResource),
+  requireResource: reduxJsonApiAliasActionToTabMarkedThunk(originalRequireResource),
 };
